refactor: move fixed-queeu to dispatcher and rm node folder
