+++ lib/dispatcher/pool-base.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const DispatcherBase = require('./dispatcher-base')
-const FixedQueue = require('../node/fixed-queue')
+const FixedQueue = require('./fixed-queue')
 const { kConnected, kSize, kRunning, kPending, kQueued, kBusy, kFree, kUrl, kClose, kDestroy, kDispatch } = require('../core/symbols')
 const PoolStats = require('./pool-stats')
